// practice/hw_ang_ngrx_p2/My-APM/src/app/products/interfaces/product.interface.ts
import * as fromRoot from '../../state/app.state';
import {Product} from '../product';

export interface State extends fromRoot.State {
  products: ProductState;
}

export interface ProductState {
  showProductCode: boolean;
  currentProductId: number | null;
  products: Product[];
  error: string;
}